Sketching out DN classes & changing all DN floats to doubles

+++ NXTAPE/src/org/opendrawer/dawinian/neurodynamics/DataProvider.java
@@ -3,7 +3,7 @@ package org.opendrawer.dawinian.neurodynamics;
 public interface DataProvider {
 	public abstract String getName();
 
-	public abstract float[] getNormalizedValues();
+	public abstract double[] getNormalizedValues();
 
 	public abstract String[] getChannelNames();
