+++ NXTAPE/src/org/opendrawer/dawinian/neurodynamics/InputProvider.java
@@ -3,5 +3,5 @@ package org.opendrawer.dawinian.neurodynamics;
 public interface InputProvider extends DataProvider {
 	public abstract int getInputChannelCount();
 
-	public abstract void setInputChannels(float[] data);
+	public abstract void setInputChannels(double[] data);
 }
